moved constant_pad to raster utils
